requirements.txt
python-dotenv
langchain-core
langchain-community
langchain-openai
langchain-groq
rdkit
pillow
streamlit
